news_collection/lambda_3_generateImages/requirements.txt
boto3==1.35.26
botocore==1.35.26
certifi==2024.8.30
cloudinary==1.41.0
DateTime==5.5
grpcio==1.63.2
grpcio-tools==1.63.2
jmespath==1.0.1
numpy==2.1.1
pandas==2.2.3
param==2.1.1
pillow==10.4.0
protobuf==5.28.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
s3transfer==0.10.2
setuptools==75.1.0
six==1.16.0
stability-sdk==0.8.6
tzdata==2024.2
urllib3==2.2.3
zope.interface==7.0.3
